Revert change of exported functions
